Refactor: Move all classes to Model folder and update logic accordingly
- Moved all C# class files (Course, Student, Diploma, Requirement, Repository, etc.) into the 'Model' folder to improve project structure and organization.
- Updated namespaces and references in the code to reflect the new folder structure.
- Ensured all unit tests work correctly with the updated structure.
- Improved code readability and maintainability.
