+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,7 @@
+{
+  "ExpandedNodes": [
+    ""
+  ],
+  "SelectedNode": "\\GraduationTracker.sln",
+  "PreviewInSolutionExplorer": false
+}
