+++ GraduationTracker/GraduationTracker.Tests.Unit/GraduationTrackerTests.cs
@@ -2,6 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using System.Linq;
+using GraduationTracker.Model;
 
 namespace GraduationTracker.Tests.Unit
 {
@@ -13,80 +14,23 @@ namespace GraduationTracker.Tests.Unit
         [TestMethod]
         public void TestEligibilityForAllStudents()
         {
-            // Arrange: create an instance of the tracker and the diploma definition
+            // Arrange
             var tracker = new GraduationTracker();
 
-            var diploma = new Diploma
+            // Get diploma and students from Repository (use the IDs accordingly)
+            var diploma = Repository.GetDiploma(1);
+
+            // Use all students defined in the repository
+            var students = new Student[]
             {
-                Id = 1,
-                Credits = 4,
-                Requirements = new int[] { 100, 102, 103, 104 }
+                Repository.GetStudent(1),
+                Repository.GetStudent(2),
+                Repository.GetStudent(3),
+                Repository.GetStudent(4),
+                Repository.GetStudent(5),
+                Repository.GetStudent(6)
             };
 
-            var students = new[]
-            {
-               new Student
-               {
-                   Id = 1,
-                   Courses = new Course[]
-                   {
-                        new Course{Id = 1, Name = "Math", Mark=95 },
-                        new Course{Id = 2, Name = "Science", Mark=95 },
-                        new Course{Id = 3, Name = "Literature", Mark=95 },
-                        new Course{Id = 4, Name = "Physichal Education", Mark=95 }
-                   }
-               },
-               new Student
-               {
-                   Id = 2,
-                   Courses = new Course[]
-                   {
-                        new Course{Id = 1, Name = "Math", Mark=80 },
-                        new Course{Id = 2, Name = "Science", Mark=80 },
-                        new Course{Id = 3, Name = "Literature", Mark=80 },
-                        new Course{Id = 4, Name = "Physichal Education", Mark=80 }
-                   }
-               },
-            new Student
-            {
-                Id = 3,
-                Courses = new Course[]
-                {
-                    new Course{Id = 1, Name = "Math", Mark=50 },
-                    new Course{Id = 2, Name = "Science", Mark=50 },
-                    new Course{Id = 3, Name = "Literature", Mark=50 },
-                    new Course{Id = 4, Name = "Physichal Education", Mark=50 }
-                }
-            },
-            new Student
-            {
-                Id = 4,
-                Courses = new Course[]
-                {
-                    new Course{Id = 1, Name = "Math", Mark=40 },
-                    new Course{Id = 2, Name = "Science", Mark=40 },
-                    new Course{Id = 3, Name = "Literature", Mark=40 },
-                    new Course{Id = 4, Name = "Physichal Education", Mark=40 }
-                }
-            },
-             // Student 5: No courses – should be marked as Remedial
-             new Student
-            {
-                Id = 5,
-                Courses = new Course[0]
-            },
-              // Student 6: Missing required course (Literature) – should not graduate
-             new Student
-            {
-                Id = 6,
-                Courses = new Course[]
-                {
-                    new Course{Id = 1, Name = "Math", Mark=90},
-                    new Course{Id = 2, Name = "Science", Mark=90},
-                    new Course{Id = 4, Name = "Physichal Education", Mark=90} // Missing Literature
-                }
-            }
-        };
             var testCases = new[]
             {
                 (Student: students[0], ExpectedGraduated: true, ExpectedStanding: STANDING.SumaCumLaude),
