+++ GraduationTracker/GraduationTracker/GraduationTracker.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using GraduationTracker.Model;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
