+++ GraduationTracker/GraduationTracker/Model/Diploma.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GraduationTracker
+namespace GraduationTracker.Model
 {
     public class Diploma
     {
