+++ GraduationTracker/GraduationTracker/Model/Requirement.cs
@@ -1,4 +1,4 @@
-﻿namespace GraduationTracker
+﻿namespace GraduationTracker.Model
 {
     public class Requirement
     {
