+++ GraduationTracker/GraduationTracker/Model/STANDING.cs
@@ -1,4 +1,4 @@
-﻿namespace GraduationTracker
+﻿namespace GraduationTracker.Model
 {
     public enum STANDING
     {
